aiml/aiLLMs/apps/backend/requirements.txt
fastapi
uvicorn
pillow
scikit-learn
pandas
numpy
tqdm
python-multipart
opencv-python
matplotlib